+++ pynetbox_query/pynetbox_query/utils/device_dataclass.py
